requirements.txt
streamlit>=1.30.0
pypdf>=4.0.0
python-docx>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
plotly>=5.18.0
